检验计划接口修改

+++ intelligent-warehouse/src/main/java/org/jeecg/modules/warehouse/service/impl/CheckPlanServiceImpl.java
@@ -5,6 +5,7 @@ import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang.StringUtils;
+import org.jeecg.common.api.vo.Result;
 import org.jeecg.modules.datasources.core.cron.CronExpression;
 import org.jeecg.modules.datasources.mapper.WaterfallJobInfoMapper;
 import org.jeecg.modules.datasources.mapper.WaterfallJobLogMapper;
@@ -15,6 +16,7 @@ import org.jeecg.modules.warehouse.dto.WaterfallQualityCheckPlanDTO;
 
 import org.jeecg.modules.warehouse.mapper.*;
 import org.jeecg.modules.warehouse.model.WaterfallQualityModelWithJobInfo;
+import org.jeecg.modules.warehouse.model.WaterfallQualityRule;
 import org.jeecg.modules.warehouse.model.WaterfallQualityRuleField;
 import org.jeecg.modules.warehouse.model.WaterfallQualityRuleWithJobInfo;
 import org.jeecg.modules.warehouse.service.ICheckPlanService;
@@ -23,7 +25,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 /**
  * @Author liansongye
@@ -73,7 +78,17 @@ public class CheckPlanServiceImpl implements ICheckPlanService {
 
     @Override
     public void deleteCheckPlan(Integer id) {
-        modelWithJobInfoMapper.deleteByPrimaryKey(id);
+        WaterfallQualityModelWithJobInfo qualityModelWithJobInfo = modelWithJobInfoMapper.selectByPrimaryKey(id);
+        if (qualityModelWithJobInfo != null) {
+            modelWithJobInfoMapper.deleteByPrimaryKey(id);
+            ruleWithJobInfoMapper.delete(new LambdaQueryWrapper<WaterfallQualityRuleWithJobInfo>()
+                    .eq(WaterfallQualityRuleWithJobInfo::getJobInfoId, qualityModelWithJobInfo.getJobInfoId()));
+        }
+    }
+
+    @Override
+    public List<WaterfallQualityRule> queryRuleList(Integer modelId, Integer jobId) {
+        return ruleMapper.selectListWithSelect(modelId, jobId);
     }
 
 
@@ -98,7 +113,7 @@ public class CheckPlanServiceImpl implements ICheckPlanService {
 
         jobInfo.setExecutorRouteStrategy("RANDOM");
 //        jobInfo.setExecutorBlockStrategy("SERIAL_EXECUTION");
-        jobInfo.setExecutorHandler("checkPlanJobHandler");
+        jobInfo.setExecutorHandler("checkPlanJobHandler2");
         jobInfo.setGlueUpdatetime(new Date());
         waterfallJobInfoMapper.insert(jobInfo);
 
@@ -127,9 +142,9 @@ public class CheckPlanServiceImpl implements ICheckPlanService {
         ruleFields.stream().forEach(e -> {
             if (StringUtils.isNotBlank(e.getRegularExpression())) {
                 dbInfo.getCornCheck().put(e.getFieldName(), e.getRegularExpression());
-            }else if (e.getEmptyFlag() != null && e.getEmptyFlag() == true) {
+            } else if (e.getEmptyFlag() != null && e.getEmptyFlag() == true) {
                 dbInfo.getEmptyCheck().add(e.getFieldName());
-            }else if (StringUtils.isNotBlank(e.getCompareMode()) && StringUtils.isNotBlank(e.getExpectedValue())) {
+            } else if (StringUtils.isNotBlank(e.getCompareMode()) && StringUtils.isNotBlank(e.getExpectedValue())) {
                 dbInfo.getCompareCheck().put(e.getFieldName(), e.getCompareMode() + e.getExpectedValue());
             }
         });
@@ -138,10 +153,14 @@ public class CheckPlanServiceImpl implements ICheckPlanService {
 
     @Override
     public WaterfallQualityCheckPlanDTO checkPlanInfo(Integer jobId) {
-         WaterfallQualityCheckPlanDTO res = modelWithJobInfoMapper.checkPlanInfo(jobId);
-         if (res != null) {
-             res.setQualityRules(ruleMapper.getListWithJob(jobId));
-         }
+        WaterfallQualityCheckPlanDTO res = modelWithJobInfoMapper.checkPlanInfo(jobId);
+        if (res != null) {
+//            res.setQualityRules(ruleMapper.getListWithJob(jobId));
+            List<WaterfallQualityRule> waterfallQualityRules = queryRuleList(res.getModelId(), jobId);
+            List<Integer> hasSelect = waterfallQualityRules.stream().filter(e -> e.getHasSelect() != null).map(e -> e.getId()).collect(Collectors.toList());
+            res.setQualityRules(waterfallQualityRules);
+            res.setHasSelect(hasSelect);
+        }
         return res;
     }
